eval for guetzli

+++ eval/start_aflpp.sh
@@ -9,19 +9,22 @@ trap _term SIGINT
 
 set -Euxo pipefail
 
-rm -rf input
-mkdir input
-echo "a" > input/0
+pwd
 
-rm -rf output
+afl-clang-fast++ $1 $2
+
+[[ -d output ]] && rm -rf output
 mkdir output
 
-afl-clang-lto $1
+shift
+shift
+
+SEEDS="$1"
 
 shift
 
 export AFL_NO_AFFINITY=1
-afl-fuzz -d -i input -o output -- ./a.out $@ &
+afl-fuzz -d -i $SEEDS -o output -- ./a.out $@ &
 
 child=$! 
 wait "$child"
